% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_vasoactive_col.R
\name{create_vasoactive_col}
\alias{create_vasoactive_col}
\title{Add a column to indicate if other infusion that will interact
with blood pressure is in use}
\usage{
create_vasoactive_col(df)
}
\description{
Add a column to indicate if other infusion that will interact
with blood pressure is in use
}
