% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_rsplit.R
\name{custom_rsplit}
\alias{custom_rsplit}
\title{Make rsplit object}
\usage{
custom_rsplit(data, in_id, out_id)
}
\arguments{
\item{in_id}{vector of integers taken from data_split where
the column \code{train} equals 1 (to select training set)}

\item{out_id}{vector of integers taken from data_split where
the column \code{train} equals 0 (to select testing set)}
}
\description{
This is a function from rsample that is
not exported but we need it because we use our own method
for the split that takes into account patient id
}
