% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_age.R
\name{format_age}
\alias{format_age}
\title{Format age variable}
\usage{
format_age(df)
}
\arguments{
\item{df}{dataframe modified in make_data_format}
}
\description{
Converts age variable from character string with
over 89 to 90 as a numeric
}
