% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_count_nitro_infusion.R
\name{format_count_nitro_infusion}
\alias{format_count_nitro_infusion}
\title{Count number of nitro dose titrations}
\usage{
format_count_nitro_infusion(df)
}
\arguments{
\item{df}{dataframe modified in make_data_format}
}
\value{
dataframe to be further modified in make_data_format
}
\description{
Adds the following variables:
\itemize{
\item \code{n_nitro}, which enumerates the doses of nitro within levels of 'id'
\item \code{time_until_nitro}, which is the time until the next dose of nitro
\item \code{time_since_nitro}, which is the time since the last dose of nitro
}
}
