% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_fill.R
\name{format_fill}
\alias{format_fill}
\title{Fill intermediate rows with infusion dose or lab results}
\usage{
format_fill(df)
}
\arguments{
\item{df}{dataframe modified in format_raw function}
}
\description{
Fill intermediate rows with infusion dose or lab results
}
