% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_height.R
\name{format_height}
\alias{format_height}
\title{Format height coloumn}
\usage{
format_height(df)
}
\arguments{
\item{df}{dataframe modified in make_data_format}
}
\description{
Changes name of column to height
}
