% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_infusions.R
\name{format_infusions}
\alias{format_infusions}
\title{format data for infusions}
\usage{
format_infusions(df)
}
\arguments{
\item{df}{data frame produced from make_infusions}
}
\description{
This function will create a variable called \code{nitro}
which is the dose rate
}
