% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_nitro_post.R
\name{format_nitro_post}
\alias{format_nitro_post}
\title{Format nitro dose column}
\usage{
format_nitro_post(df)
}
\arguments{
\item{df}{dataframe modified in make_data_format}
}
\description{
Changes name of column to nitro_post
}
