% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_sbp_post.R
\name{format_sbp_post}
\alias{format_sbp_post}
\title{Add details of dose and time after dose titration that
blood pressure was taken}
\usage{
format_sbp_post(df)
}
\arguments{
\item{df}{dataframe created in \code{make_data_frame}}
}
\description{
Adds:
\itemize{
\item \code{sbp_post}, which is the most recent value of
sbp before next dose of nitro
\item \code{time_post}, which is the time that sbp_post occured,
in minutes after dose of nitro
}
}
