% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_sbp_pre.R
\name{format_sbp_pre}
\alias{format_sbp_pre}
\title{Add details of dose and time before dose titration that
blood pressure was taken}
\usage{
format_sbp_pre(df)
}
\arguments{
\item{df}{dataframe created in \code{make_data_frame}}
}
\description{
Adds:
\itemize{
\item \code{sbp_pre}, which is the most recent value of
sbp before next dose of nitro
\item \code{time_pre}, which is the time that sbp_pre occured,
in minutes before next dose of nitro
\item \code{time_since_nitro}, which is the time since the last
dose of nitro
}
}
