% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_tables.R
\name{join_tables}
\alias{join_tables}
\title{Combine data from tables and format}
\usage{
join_tables(
  patient,
  nitro,
  sbp_nurse_charting,
  vital_aperiodic,
  vital_periodic,
  apachepatientresult,
  apachepredvar,
  creat,
  fentanyl,
  midaz,
  propofol,
  dobutamine,
  dex,
  nicardipine,
  amiodarone,
  fluids,
  milrinone,
  epinephrine,
  vasopressin,
  diltiazem,
  pain
)
}
\arguments{
\item{patient}{dataframe from SQL query}

\item{nitro}{dataframe from SQL query}

\item{sbp_nurse_charting}{dataframe from SQL query}

\item{vital_aperiodic}{dataframe from SQL query}

\item{vital_periodic}{dataframe from SQL query}

\item{apachepatientresult}{dataframe from SQL query}

\item{creat}{dataframe from SQL query}
}
\value{
target is joinedTables
}
\description{
Combine data from different tables and format
}
