% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_invasive.R
\name{make_invasive}
\alias{make_invasive}
\title{invasive sbp data}
\usage{
make_invasive(data_formatted)
}
\arguments{
\item{data_formatted}{target from pipeline}
}
\value{
data_invasive pipeline target
}
\description{
Invasive sbp measurements for patients who have
had at least 5 nitro dose titrations (n=246)
}
