% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sbp.R
\name{make_sbp}
\alias{make_sbp}
\title{make_sbp}
\usage{
make_sbp(df)
}
\arguments{
\item{df}{data frame produced from make_infusions}
}
\value{

}
\description{
This function will create a variable called \code{sbp}
that takes the value of either the non-invasive or systemic
blood pressure from the vitalaperiodic, vitalperiodic and
nursecharting tables. It preferentially takes the systemic
measurement when both systemic and non-invasive is available.
}
