% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sbp_measure.R
\name{make_sbp_measure}
\alias{make_sbp_measure}
\title{Split invasive and non-invasive sbp measurements}
\usage{
make_sbp_measure(df)
}
\arguments{
\item{df}{dataframe in make_data_format}
}
\value{
dataframe with column to indicate type of sbp measurement
}
\description{
Split invasive and non-invasive sbp measurements
}
