% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_total_nitro.R
\name{make_total_nitro}
\alias{make_total_nitro}
\title{Add total dose of nitro}
\usage{
make_total_nitro(df)
}
\arguments{
\item{df}{data_formatted from the pipeline}
}
\description{
This function will add a column that is a
total dose of nitroglycerin the patient has received.
The dose is reset when the infusion dose is zero.
}
