% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nitro_dose_debug.R
\name{nitro_dose_debug}
\alias{nitro_dose_debug}
\title{Remove unrealistic nitro doses}
\usage{
nitro_dose_debug(df)
}
\arguments{
\item{df}{data_formatted from the pipeline}
}
\description{
This function will remove the
rows where doses less than 1mcg/min were noted
}
