% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_cols.R
\name{query_cols}
\alias{query_cols}
\title{Download data from bigquery using only a SELECT condition}
\usage{
query_cols(connection, table, columns)
}
\arguments{
\item{connection}{DBI connection to bigquery}

\item{table}{name of table to query}

\item{columns}{vector of column names from the table to select}

\item{query}{SQL query as a string}
}
\value{
Character string of SQL query
}
\description{
Queries run using this function select
all the entries from the tables for all patients
who had at least one entry for nitroglycerin in the
infusiondrug table
}
