% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_rows.R
\name{query_rows}
\alias{query_rows}
\title{Download data from bigquery using both SELECT and
WHERE conditions}
\usage{
query_rows(connection, table, columns, rows)
}
\arguments{
\item{connection}{DBI connection to bigquery}

\item{table}{name of table to query}

\item{columns}{vector of column names from the table to select}

\item{rows}{condition to pass to WHERE query for the table}
}
\value{
dataframe with data from table in eicu dataset
stored in bigquery
}
\description{
Queries run using this function select
all the entries from the tables for all patients
who had at least one entry for nitroglycerin in the
infusiondrug table
}
