% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_connection.R
\name{make_connection}
\alias{make_connection}
\title{Connect with database}
\usage{
make_connection(billing, dataset)
}
\arguments{
\item{billing}{Account for billing in bigquery}

\item{dataset}{name of bigquery dataset where
eicu data has been copied into}
}
\value{
Connection to eicu database from
your copy of physionet-data in bigquery
}
\description{
Connect with database
}
