% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_format.R
\name{make_data_format}
\alias{make_data_format}
\title{Combine data from tables and format}
\usage{
make_data_format(df, create_vasoactive_col = TRUE)
}
\arguments{
\item{df}{df returned from join_tables}

\item{create_vasoactive_col}{logial creates a column to
indicate that a patient was receiving another infusion
that is likely to influence blood pressure}
}
\value{
target is data_formatted
}
\description{
Combine data from different tables and format
}
